+++ agent/src/test/groovy/nu/studer/teamcity/buildscan/agent/gradle/BuildScanUrlCapturingInitScriptTest.groovy
@@ -4,7 +4,7 @@ import static org.junit.Assume.assumeTrue
 
 class BuildScanUrlCapturingInitScriptTest extends BaseInitScriptTest {
 
-    def "sends build started service message even without declaring Gradle Enterprise / Build Scan plugin (#jdkCompatibleGradleVersion)"() {
+    def "sends build started service message even without declaring Develocity / Build Scan plugin (#jdkCompatibleGradleVersion)"() {
         assumeTrue jdkCompatibleGradleVersion.isJvmVersionCompatible()
 
         when:
@@ -17,11 +17,11 @@ class BuildScanUrlCapturingInitScriptTest extends BaseInitScriptTest {
         jdkCompatibleGradleVersion << GRADLE_VERSIONS_3_0_AND_HIGHER
     }
 
-    def "send build scan url service message when declaring Gradle Enterprise / Build Scan plugin (#jdkCompatibleGradleVersion)"() {
+    def "send build scan url service message when declaring Develocity / Build Scan plugin (#jdkCompatibleGradleVersion)"() {
         assumeTrue jdkCompatibleGradleVersion.isJvmVersionCompatible()
 
         given:
-        declareGePluginApplication(jdkCompatibleGradleVersion.gradleVersion)
+        declareDevelocityPluginApplication(jdkCompatibleGradleVersion.gradleVersion)
 
         when:
         def result = run(jdkCompatibleGradleVersion.gradleVersion)