+++ agent/src/test/groovy/nu/studer/teamcity/buildscan/agent/maven/ServiceMessageExtensionApplicationTest.groovy
@@ -7,20 +7,20 @@ import static org.junit.Assume.assumeTrue
 
 class ServiceMessageExtensionApplicationTest extends BaseExtensionApplicationTest {
 
-    def "build succeeds when service message maven extension is applied to a project without GE in the extension classpath (#jdkCompatibleMavenVersion)"() {
+    def "build succeeds when service message maven extension is applied to a project without Develocity in the extension classpath (#jdkCompatibleMavenVersion)"() {
         assumeTrue jdkCompatibleMavenVersion.isJvmVersionCompatible()
-        assumeTrue GE_URL != null
+        assumeTrue DEVELOCITY_URL != null
 
         given:
         def mvnProject = new MavenProject.Configuration().buildIn(checkoutDir)
 
         and:
-        def gePluginConfig = new TcPluginConfig(
+        def develocityPluginConfig = new TcPluginConfig(
             enableCommandLineRunner: true,
         )
 
         when:
-        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, gePluginConfig)
+        def output = run(jdkCompatibleMavenVersion.mavenVersion, mvnProject, develocityPluginConfig)
 
         then:
         outputContainsBuildSuccess(output)