+++ src/test/groovy/nu/studer/teamcity/buildscan/connection/DevelocityPasswordProviderTest.groovy
@@ -1,21 +1,20 @@
 package nu.studer.teamcity.buildscan.connection
 
 import jetbrains.buildServer.parameters.ParametersProvider
-import jetbrains.buildServer.serverSide.Parameter
 import jetbrains.buildServer.serverSide.SBuild
 import jetbrains.buildServer.serverSide.parameters.types.PasswordsProvider
 import spock.lang.Specification
 
-import static nu.studer.teamcity.buildscan.connection.GradleEnterpriseConnectionConstants.GRADLE_ENTERPRISE_ACCESS_KEY_ENV_VAR
+import static DevelocityConnectionConstants.GRADLE_ENTERPRISE_ACCESS_KEY_ENV_VAR
 
-class GradleEnterprisePasswordProviderTest extends Specification {
+class DevelocityPasswordProviderTest extends Specification {
 
     PasswordsProvider passwordsProvider
     SBuild sBuild
     ParametersProvider parametersProvider
 
     void setup() {
-        passwordsProvider = new GradleEnterprisePasswordProvider()
+        passwordsProvider = new DevelocityPasswordProvider()
 
         sBuild = Stub(SBuild)
         parametersProvider = Stub(ParametersProvider)
@@ -35,7 +34,7 @@ class GradleEnterprisePasswordProviderTest extends Specification {
 
     def "returns env.GRADLE_ENTERPRISE_ACCESS_KEY element when env.GRADLE_ENTERPRISE_ACCESS_KEY is set"() {
         given:
-        def value = "ge.example.com=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+        def value = "develocity.example.com=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
         parametersProvider.get(GRADLE_ENTERPRISE_ACCESS_KEY_ENV_VAR) >> value
 
         when: