+++ release/changes.md
@@ -1,1 +1,1 @@
-- Informs the Gradle Enterprise Gradle plugin that it is being applied externally.
+- Informs the Develocity Gradle plugin that it is being applied externally.